Add wdr.py
Adds wdr.py, which is just an improved wdr-read.py. Thanks to utopiadeferred, we're able to deduce that to obtain the vertexdata offset, we need to read the vertex buffer & the dataoffset to vertexdata, followed by jumping + physical file size + stride. As well - there are multiple strides that are possible for vertex data, with 28 and 68 being the rarest.

- Thanks Utopia for the help!